+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Fixed:
 - Don't measure the height of `RedwoodUIView` as zero when measured with `sizeThatFits()` or `intrinsicContentSize()`. We had been using `UIStackView` which doesn't support these functions.
 - Correctly signal `RedwoodUIView` size changes of to callers who measure it with `intrinsicContentSize()`.
 - The Redwood Gradle plugin is now compatible with Gradle 9.1.
+- Don't incorrectly size items to 0 when using the `Flex()` modifier on a `Wrap` container.
 
 
 ## [0.18.0] - 2025-08-01